+++ symbol_by_market_phase.py
@@ -1,4 +1,5 @@
 from bec.symbol_by_market_phase import *
 
 if __name__ == "__main__":
-    raise SystemExit(cli_main())
+    time_frame, trade_against_value = read_arguments()
+    main(timeframe=time_frame)